+++ src/e_mod_main.h
@@ -74,7 +74,7 @@ struct _Config_Item
    const char *area_text_5; 
    /* actual config properties independently for every running gadget. */
    int header_switch, multiply_switch, notif_switch;
-   double font_size;
+   int font_size;
    double interval;
    struct
    {
